+++ sdk/ml/azure-ai-ml/azure/ai/ml/_schema/core/fields.py
@@ -657,6 +657,28 @@ def DistributionField(**kwargs):
     )
 
 
+def PrimitiveValueField(**kwargs):
+    return UnionField(
+        [
+            # Note: order matters here - to make sure value parsed correctly.
+            # By default when strict is false, marshmallow downcasts float to int.
+            # Setting it to true will throw a validation error when loading a float to int.
+            # https://github.com/marshmallow-code/marshmallow/pull/755
+            # Use DumpableIntegerField to make sure there will be validation error when
+            # loading/dumping a float to int.
+            # note that this field can serialize bool instance but cannot deserialize bool instance.
+            DumpableIntegerField(strict=True),
+            # Use DumpableFloatField with strict of True to avoid '1'(str) serialized to 1.0(float)
+            DumpableFloatField(strict=True),
+            # put string schema after Int and Float to make sure they won't dump to string
+            fields.Str(),
+            # fields.Bool comes last since it'll parse anything non-falsy to True
+            fields.Bool(),
+        ],
+        **kwargs,
+    )
+
+
 class VersionField(Field):
     def __init__(self, *args, **kwargs) -> None:
         super().__init__(*args, **kwargs)