+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/_text_analytics_client.py
@@ -23,7 +23,6 @@ from ._request_handlers import (
     _determine_action_type,
 )
 from ._validate import validate_multiapi_args, check_for_unsupported_actions_types
-from ._version import DEFAULT_API_VERSION
 from ._response_handlers import (
     process_http_response_error,
     entities_result,
@@ -145,7 +144,6 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
         super().__init__(
             endpoint=endpoint, credential=credential, **kwargs
         )
-        self._api_version = kwargs.get("api_version", DEFAULT_API_VERSION)
         self._default_language = kwargs.pop("default_language", "en")
         self._default_country_hint = kwargs.pop("default_country_hint", "US")
         self._string_index_type_default = (
@@ -658,7 +656,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
             2022-10-01-preview and newer). If not set, uses "en" for English as default.
             Per-document language will take precedence over whole batch language.
             See https://aka.ms/talangs for supported languages in Language API.
-        :keyword autodetect_default_language: Default/fallback language to use for documents requesting
+        :keyword str autodetect_default_language: Default/fallback language to use for documents requesting
             automatic language detection.
         :keyword str display_name: An optional display name to set for the requested analysis.
         :keyword str string_index_type: Specifies the method used to interpret string offsets.
@@ -1121,6 +1119,9 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
         combine multiple text analysis actions into one call. Otherwise, we recommend you use
         the action specific endpoints, for example :func:`analyze_sentiment`.
 
+        .. note:: The abstractive summarization feature is part of a gated preview. Request access here:
+            https://aka.ms/applyforgatedsummarizationfeatures
+
         .. note:: See the service documentation for regional support of custom action features:
             https://aka.ms/azsdk/textanalytics/customfunctionalities
 
@@ -1147,7 +1148,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
             2022-10-01-preview and newer). If not set, uses "en" for English as default.
             Per-document language will take precedence over whole batch language.
             See https://aka.ms/talangs for supported languages in Language API.
-        :keyword autodetect_default_language: Default/fallback language to use for documents requesting
+        :keyword str autodetect_default_language: Default/fallback language to use for documents requesting
             automatic language detection.
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword int polling_interval: Waiting time between two polls for LRO operations
@@ -1368,7 +1369,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
             2022-10-01-preview and newer). If not set, uses "en" for English as default.
             Per-document language will take precedence over whole batch language.
             See https://aka.ms/talangs for supported languages in Language API.
-        :keyword autodetect_default_language: Default/fallback language to use for documents requesting
+        :keyword str autodetect_default_language: Default/fallback language to use for documents requesting
             automatic language detection.
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword bool disable_service_logs: If set to true, you opt-out of having your text input
@@ -1494,7 +1495,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
             2022-10-01-preview and newer). If not set, uses "en" for English as default.
             Per-document language will take precedence over whole batch language.
             See https://aka.ms/talangs for supported languages in Language API.
-        :keyword autodetect_default_language: Default/fallback language to use for documents requesting
+        :keyword str autodetect_default_language: Default/fallback language to use for documents requesting
             automatic language detection.
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword bool disable_service_logs: If set to true, you opt-out of having your text input
@@ -1614,7 +1615,7 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
             2022-10-01-preview and newer). If not set, uses "en" for English as default.
             Per-document language will take precedence over whole batch language.
             See https://aka.ms/talangs for supported languages in Language API.
-        :keyword autodetect_default_language: Default/fallback language to use for documents requesting
+        :keyword str autodetect_default_language: Default/fallback language to use for documents requesting
             automatic language detection.
         :keyword bool show_stats: If set to true, response will contain document level statistics.
         :keyword bool disable_service_logs: If set to true, you opt-out of having your text input
@@ -1715,6 +1716,9 @@ class TextAnalyticsClient(TextAnalyticsClientBase):
         On the fly classification of the input documents into one or multiple categories.
         Assigns either one or multiple categories per document. This type of classification
         doesn't require model training.
+
+        .. note:: The dynamic classification feature is part of a gated preview. Request access here:
+            https://aka.ms/applyforgatedlanguagefeature
 
         See https://aka.ms/azsdk/textanalytics/data-limits for service data limits.