+++ sdk/quota/azure-mgmt-quota/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release History
 
+## 1.1.0b3 (2022-11-09)
+
+### Other Changes
+
+  - Added generated samples in github repo
+  - Drop support for python<3.7.0
+
 ## 1.0.0b2 (2021-11-01)
 
 **Features**